# Optimize startup time using lazy imports