Fixes #514 

The goal of this PR is to reduce CLI startup time by avoiding useless work, especially imports that are not necessary for the requested operation.

It makes the following changes to the import statements within the Annif codebase:

- complete rewrite of `annif/backend/__init__.py`; the end result is that backends (and the libraries they require, e.g. fasttext, omikuji and tensorflow) are only imported when they are actually used
- avoid importing NLTK, sklearn, joblib and NumPy unless actually required, by moving import statements inside functions and methods

I tried to craft the changes to have minimal impact on the code so I only chose to make imports local in cases where there were very few uses within the same module. For example `annif.suggestion` uses NumPy a lot, so I didn't try to use local imports within that module, but instead avoided importing the whole `annif.suggestion` module in other parts of the codebase.

Startup time for simple commands such as `annif --help` and `annif --version` has been reduced by more than two thirds.

Before:

```
$ time annif --version
0.56.0.dev0

real	0m4,052s
user	0m4,001s
sys	0m0,568s
```

After:

```
$ time annif --version
0.56.0.dev0

real	0m1,302s
user	0m1,250s
sys	0m0,050s
```

As explained in #514, I also used [tuna](https://github.com/nschloe/tuna) to visualize where the remaining import time is spent after this PR:

![image](https://user-images.githubusercontent.com/1132830/146564388-f02a4544-740b-45ad-a1e0-afbf795eabf0.png)

The main culprits are now `connexion` (with most of the time spent initializing `openapi_spec_validator`!) and `flask`. Those are core libraries and I don't think we can avoid importing them even for the simplest CLI commands.