@osma 2021-12-17
I also looked at the option of implementing truly lazy imports, for example as [described](https://docs.python.org/3/library/importlib.html#implementing-lazy-imports) in the importlib standard module. But that didn't seem very appealing - it takes a nontrivial amount of code and can be hard to debug if the imports fail for some reason. So I chose to use local imports instead, which are easier to understand if not always so elegant.