@osma 2021-12-17
Looking at CI test results and especially Codecov, I think we need additional tests for the ImportError/ValueError clauses in `annif/backend/__init__.py`