@osma 2021-12-20
I will try again with a less invasive approach that doesn't try to avoid importing NumPy. It's not worth complicating the code with many local imports just to speed up the startup by less than 0.1 seconds.